requirements.txt
Flask==2.0.1
requests==2.26.0
Pillow==8.3.1
python-dotenv==0.19.0